dragonfs: increase chances to fastpath in dfs_read
In dfs_read, there's a fastpath that bypasses the internal file cache
and DMA the data directly into the output buffer. The existing code
assumed that this was possible only when both the RDRAM and ROM addresses
and the size were all multiple of 8.

This is wrong: PI is able to DMA from any even ROM address; also, the DMA
length can be of any size, including odd numbers. The constraint of
8-bytes aligned only stands for RDRAM.

This commit also modifies the testsuite to make sure all cases are
excercised.