New RSP DMA code library
This commit adds rsp_dma.inc, a simple yet powerful piece of code
to run DMA transfers from within the RSP. It is packaged as small code
library that is sufficient to include at any point in the text segment
to compile in.

This code comes from multiple years of evolution in writing DMA
functions in RSP, and it is a good compromise between features and code
size which I feel it can be used as default for all RSP microcodes in
libdragon.

It supports both synchronous and asynchronous (background) transfers,
and it allows to maximize RSP DMA usage by enqueuing a transfer while
another one is in progress.