New XM64 player.
Based on the previous commits, this commit adds the XM64 player.
The implementation is remarkably easy (see tick()), as the mixer
provides a very easy API to do sequencing and to playback and mix
digital waveforms.

Performance is extremely good. XM64 files up to 8 channels are usually
played back using a few percents of CPU, and less than 10% of RSP time,
with together with module compactness makes the format really a good
choice for background music in games, without removing too many resources
from the graphics.