cop0: use volatile in return values of hardware reads
This is a quick workaround to a general problem with inline assembly
where it's basically impossible to model the fact that a value read
from a hardware register can vary unexpectedly and moving other operation
before/after the inline assembly can change the meaning of the program.
This seems to fix most instances I noticed where TICKS_READ (aka C0_COUNT)
was being generated in the wrong position because the compiler thought
that something after it could be done before (basically invalidating
the profiling being done).