Refactor RSP library.
This is a refactoring of the existing library. Functions are renamed
to provide a fixed prefix for namespacing concerns (old names are
still available and marked as deprecated).

rsp.h now exposes some hardware register definitions so that advanced
users can play with RSP without having to redefine them. This is
in line with the trend of letting libdragon also expose more of the
hardware without keeping it hidden within libraries.

The module is still quite barebone. There is no support for having a
queue of different ucodes (maybe switched under interrupt?) or doing
some more advanced synchronization. I think this has to wait further work
to define how RSP display lists and overlays will work. For now, this
is just a basic improvement over the existing library.