mixer: allow to playback waveforms with more then 0x40000 samples.
Currently, mixer is limited to short waveforms because the RSP uses
internally a 1.19.12 fixed point format to do resampling, so you can't
represent a byte position bigger than 2^19 (which is 2^18 for 16-bit
samples, and even 2^17 for future support of stereo waveforms). That's
too little.

This PR changes the mixer to keep internally 64-bit fixed pointer
numbers (1.51.12), and then truncate them properly when passing them
to RSP. RSP in fact doesn't need to compute across the whole waveform
but can work only with a smaller window.