New audioplayer example
This is a small audioplayer within builtin filesystem browser, for now
fully target to XM64 files. It allows to browse contents of the
filesystem, select a song for play, and display statistics during
playback, including CPU/DMA/RSP usage. It also allows interactively
to mute/demute the single mixer channels.