New YM64 player
This commit introduces a new YM64 player, which is based on the .YM
module format, a format first popularized in the Atari ST emulator scene.

The format is based around the very popular AY-3-8910 sound chip, that
was powering a few successful 8-bit consoles such as Atari ST, ZX Spectrum
and Amstrad CPC. It is a 3-channel PSG with envelope support. It can produce
typical 8-bit "chiptune" music scores. Nowadays, it is possible to compose
soundtracks using the Arkos Tracker 2 tool, that exports in the YM format.

The YM format is a simple dump of the state of all registers of the AY
chip at a fixed time step. To playback, it is necessary to emulate the
AY PSG. The implementation has been carefully optimized for the N64 MIPS
CPU for high-performance, so that playback typically takes less than 5%
of CPU time, plus a few percents of RSP time for resampling and mixing
(done by the mixer).