inthandler: process multiple interrupts in the same exception
Currently, the interrupt handler will go through the different bits in
the cause register and process the first one that is found set. If
there are multiple pending interrupts, each exception will thus process
only the first one, and then a new exception will immediately triggers.

This is wasteful: once an exception is triggered, we can as well process
all pending interrupts.