mksprite: rewrite and add support for multiple textue formats
Currently, mksprite only supports RGB16 and RGB32. This is quite
limiting. To improve RDP support in libdragon, we need to be able
to convert sprites/textures with all the supported N64 pixel formats.

This commit rewrites mksprite from scratch. First, it introduces a more
standard command line parsing with arguments (though the old syntax
is secretly accepted and used for backward compatibility with existing
Makefiles).

The input PNG file is now parsed using a vendored copy of lodepng,
a popular single-file PNG library. This drops the dependency on the
external libpng, which simplifies builds in general.

mksprite will try to autodetect the best possible pixel format
inspecting the input image file, unless explicitly told on the command
line. Notice that in any case it will never perform a quantization:
for instance, if passed a RGB16 image and told to emit a CI8 sprite,
it will refuse to do so. This is might be lifted in the future, as
bundling a (good enough) quantizer could surely help developers.

The sprite_t format is not changed. The field format used to be 0
and now is used to store the pixel format. Pelettes (when presents)
are stored in the data array before the actual pixels.

grpahics.c and rdp.c have been updated to upgrade already-converted
sprites for backward compatibility (using the bitdepth field to detect
the format), and currently assert on unsupported formats (anything but
RGB16 and RGB32).