audio: avoid allowing to overwrite a buffer that was currently played
The audio library has a historical bug: when a buffer is enqueued for
AI DMA, it is immediately flagged as "not full", which means that a
subsequent call to audio_write() could select it and overwrite its
contents while it is being played. This caused slight corruptions,
normally at the beginning of audio playback (consider 1 buffer is
40ms of a second, so the corruption could be a 10-20ms but is still
audible).

The workaround is a little more complex than expected because there is
no guarantee that we get exactly one interrupt for played buffer
(depending on interrupt latency, you could find 2 audio interrupts
batched into one). So we now keep a third index in the circular buffer
array (the "empty" index, in addition to read index and write index) and
update it any time we find that a buffer was consumed by AI, by
inspecting the AI status flags (which, in turn, requires to also
remember whether one or two buffers were enqueued).

With this fix, the corruption at start/end of playback is finally
solved. The commit also removes a broken workaround in the audioplayer
example: all buffers were muted at beginning and end of playback (which
somewhat worked because of lucky timing, but in turn caused the initial
part of the song to be skipped, again for race conditions that are not
worth describing here in details).