mixer: fix two undefined behaviors
The volume array was not zero-initialized so it could contain
non-zero values for turned-off channels. This shouldn't matter in
principle but it does matter if the DMEM BSS is initialized with random
values rather than 0 (which is what happens if there is another ucode
alternating on RSP).

Also avoid updating channel position if the channel is stopped. This
allows to correctly retrieve the channel position even after stop, which
might be useful sometimes.