interrupt: add disable_interrupt_when API
A common pattern in N64 programming is to access hardware resources
exclusively, with interrupts disabled. For instance, this is required
when programming any DMA or even just accessing the PI BUS.

Unfortunately, the optimal pattern to acquire such exclusive access
is not immediate: disabling interrupts before waiting means that the
code could keep the interrupts disabled for all the waiting time; on
the other hand, waiting for the resource to be free and then disabling
interrupts might cause race conditions.

The correct solution is to wait with interrupts enabled, then disable
interrupts and check again: if the resources has become busy, try
again the same process.

This commit introduces an API that implements this pattern, and then
use it across libdragon in all places where it makes sense to do it.
This hopefully reduces interrupt latency across the whole library.