audio: add new higher-level, easier audio_push API
For code that uses directly the audio.c library without going through
the mixer and is not concerned with total performance, the current API
is not very ergonomic because audio_write / audio_write_begin requires
filling a full buffer of samples. The caller has thus to arrange to
call these functions only when that exact number of samples is available,
handling caching in an external ring buffer.

This commit introduces a new audio_push API, which allows writing
any number of samples into the audio buffers. It does that by keeping
a state: when a buffer is acquired by audio_write_begin, it is saved
into a state, so that it can be filled through multiple calls, before
eventually being released via audio_write_end. This also works thanks
to the recent fix that makes sure that a buffer is not played back
until audio_write_end is actually called.

audio_push also allows to insert silence by simply passing a NULL
buffer, and even flushing the internal state (that is, releasing the
internal partial buffer, completing it with silence) by passing a NULL
buffer and 0 as number of samples.

Given that we now have a new higher-level API, it is time to clear
up the lower level API. We thus demote audio_write and audio_write_silence
to deprecated functions. Their use case is now covered by either
audio_write_begin (harder to use but potentially faster) or audio_push
(easier to use). It does not make sense to keep them around anymore
and we do not want to confuse users with too many alternatives. The
deprecation message shows the exact replacement to use.